fix: sync execution/legacy/loops infra from ag_new_release_updates_2
