+++ libnd4j/include/loops/cuda/pairwise_bool.cu
@@ -150,7 +150,13 @@ void PairWiseBoolTransform<X,Z>::executeCudaShaped(
       PAIRWISE_BOOL_OPS);
 }
 
+#ifdef SD_SPLIT_TYPE_INDEX
+#if COUNT_NARG(SD_COMMON_TYPES) > SD_SPLIT_TYPE_INDEX
+BUILD_DOUBLE_TEMPLATE( class PairWiseBoolTransform, , SD_SPLIT_TYPE_LIST, SD_BOOL_TYPES);
+#endif
+#else
 BUILD_DOUBLE_TEMPLATE( class PairWiseBoolTransform, , SD_COMMON_TYPES, SD_BOOL_TYPES);
+#endif
 
 }  // namespace pairwise_transforms
 }  // namespace functions
