+++ libnd4j/include/loops/cuda/reduce/reduce_bool.cu
@@ -515,7 +515,13 @@ SD_DEVICE void initializeShared(X* extraParams, X** sPartials, int sMemSize) {
 }
 
 // Explicit template instantiations for ReduceBoolFunction
+#ifdef SD_SPLIT_TYPE_INDEX
+#if COUNT_NARG(SD_COMMON_TYPES) > SD_SPLIT_TYPE_INDEX
+BUILD_DOUBLE_TEMPLATE(class ReduceBoolFunction, , SD_SPLIT_TYPE_LIST, SD_BOOL_TYPES)
+#endif
+#else
 BUILD_DOUBLE_TEMPLATE(class ReduceBoolFunction, , SD_COMMON_TYPES, SD_BOOL_TYPES)
+#endif
 
 }  // namespace reduce
 }  // namespace functions
