+++ libnd4j/include/loops/cuda/transform/transform_any.cu
@@ -185,7 +185,13 @@ SD_HOST void TransformAny<X, Z>::intermediateShaped(
   sd::DebugHelper::checkErrorCode(stream, "transformAny(...) cached kernel failed");
 }
 
+#ifdef SD_SPLIT_TYPE_INDEX
+#if COUNT_NARG(SD_COMMON_TYPES) > SD_SPLIT_TYPE_INDEX
+BUILD_DOUBLE_TEMPLATE( class TransformAny, , SD_SPLIT_TYPE_LIST, SD_COMMON_TYPES);
+#endif
+#else
 BUILD_DOUBLE_TEMPLATE( class TransformAny, , SD_COMMON_TYPES, SD_COMMON_TYPES);
+#endif
 
 }  // namespace transform
 }  // namespace functions
